Moved compatiblity utils
