+++ Arma3ObjectBuilder/utils_compat.py
